sonarqube checks

+++ voras-parent/dev.voras.framework.docker.controller/src/main/java/dev/voras/framework/docker/controller/RunPoll.java
@@ -115,7 +115,7 @@ public class RunPoll implements Runnable {
 				startPod(selectedRun);
 
 				if (!queuedRuns.isEmpty()) {
-					Thread.sleep(settings.getRunPollRecheck() * 1000);  //*** Slight delay to allow Docker to catch up
+					Thread.sleep((long)settings.getRunPollRecheck() * 1000l);  //*** Slight delay to allow Docker to catch up
 				} else {
 					return;
 				}
